Updated the way the Observer commands are added to the program
I added the subscribe function to the constructor of the "comandable" objects to streamline the adding of new commands to the program over time.  You can add new commands by using the commandable interface and passing the "this" command to the "MyGame" object "Subscribe" method by adding  the "this" to the instanciation  in the "MyGame" class commandable constructor.  (i.e. public HelpCommand HelpMe = new HelpCommand(this);)